Optimize spectrum gain to integer arithmetic (reduce CPU overhead)
- Changed from float32_t to uint32_t for gain squared calculation
- Eliminated float multiplication in spectrum processing loop (runs 16x per 100 frames)
- Integer math: gained = (accum[i] * gain*gain) / 10000
- More efficient on ARM Cortex-M33 without FPU optimization
- Fixes UDP queue regression (4/4 full) from added float operations

+++ src/waterfall.h
@@ -145,11 +145,12 @@ uint32_t waterfall_get_gain(void);
 
 /**
  * Get the squared gain value used internally for spectrum scaling
- * Computed as (gain/100)^2 for efficient per-bin multiplication
+ * Returns the scaled gain multiplier (scaled by 10000)
+ * Example: gain=10 returns 100 (represents 0.01 when divided by 10000)
  * 
- * @return Current gain squared value (float32_t)
+ * @return Scaled gain squared value (uint32_t)
  */
-float32_t waterfall_get_gain_squared(void);
+uint32_t waterfall_get_gain_squared(void);
 
 /**
  * Set the waterfall colormap by index
